requirements.txt
sounddevice>=0.4.5
numpy>=1.21.0
requests>=2.28.0
python-dotenv>=1.0.0
pyinstaller>=6.0.0
